Pruebas de la multiprogramacion de los hilos del kernel

+++ Kernel/libs/estructuras_kernel.h
@@ -10,7 +10,7 @@ typedef uint16_t t_puerto_programa;
 typedef uint16_t t_puerto_cpu;
 typedef uint16_t t_tamanio_mp;
 typedef uint16_t t_puerto_umv;
-typedef int t_tamanio_stack;
+typedef uint16_t t_tamanio_stack;
 typedef char* t_ip_umv;
 
 typedef uint8_t t_quantum;
